# openapi/openapi0008.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "DripRate Server API",
    "description": "API for DripRate Server application",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:13000/"
    }
  ],
  "paths": {
...
    "/api/Collections": {
      "post": {
        "tags": [
          "Collections"
        ],
        "operationId": "Collections_Create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCollectionRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCollectionRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCollectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "code": {
                          "$ref": "#/components/schemas/Collections_Create_409ErrorCode"
                        }
                      }
                    }
                  ]
                }
              },
              "text/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "code": {
                          "$ref": "#/components/schemas/Collections_Create_409ErrorCode"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "code": {
                          "$ref": "#/components/schemas/Collections_Create_400ErrorCode"
                        }
                      }
                    }
                  ]
                }
              },
              "text/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "code": {
                          "$ref": "#/components/schemas/Collections_Create_400ErrorCode"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "code": {
                          "$ref": "#/components/schemas/Collections_Create_401ErrorCode"
                        }
                      }
                    }
                  ]
                }
              },
              "text/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "code": {
                          "$ref": "#/components/schemas/Collections_Create_401ErrorCode"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/Collections/{id}": {
      "patch": {
        "tags": [
          "Collections"
        ],
        "operationId": "Collections_Update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCollectionRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCollectionRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCollectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "code": {
                          "$ref": "#/components/schemas/Collections_Update_403ErrorCode"
                        }
                      }
                    }
                  ]
                }
              },
              "text/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "code": {
                          "$ref": "#/components/schemas/Collections_Update_403ErrorCode"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "code": {
                          "$ref": "#/components/schemas/Collections_Update_404ErrorCode"
                        }
                      }
                    }
                  ]
                }
              },
              "text/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "code": {
                          "$ref": "#/components/schemas/Collections_Update_404ErrorCode"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "code": {
                          "$ref": "#/components/schemas/Collections_Update_409ErrorCode"
                        }
                      }
                    }
                  ]
                }
              },
              "text/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "code": {
                          "$ref": "#/components/schemas/Collections_Update_409ErrorCode"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "code": {
                          "$ref": "#/components/schemas/Collections_Update_400ErrorCode"
                        }
                      }
                    }
                  ]
                }
              },
              "text/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "code": {
                          "$ref": "#/components/schemas/Collections_Update_400ErrorCode"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "code": {
                          "$ref": "#/components/schemas/Collections_Update_401ErrorCode"
                        }
                      }
                    }
                  ]
                }
              },
              "text/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "code": {
                          "$ref": "#/components/schemas/Collections_Update_401ErrorCode"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      },
      "get": {
        "tags": [
          "Collections"
        ],
        "operationId": "Collections_GetItems",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "take",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PublicationResponse"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PublicationResponse"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PublicationResponse"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "code": {
                          "$ref": "#/components/schemas/Collections_GetItems_404ErrorCode"
                        }
                      }
                    }
                  ]
                }
              },
              "text/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "code": {
                          "$ref": "#/components/schemas/Collections_GetItems_404ErrorCode"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "code": {
                          "$ref": "#/components/schemas/Collections_GetItems_403ErrorCode"
                        }
                      }
                    }
                  ]
                }
              },
              "text/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "code": {
                          "$ref": "#/components/schemas/Collections_GetItems_403ErrorCode"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "code": {
                          "$ref": "#/components/schemas/Collections_GetItems_400ErrorCode"
                        }
                      }
                    }
                  ]
                }
              },
              "text/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "code": {
                          "$ref": "#/components/schemas/Collections_GetItems_400ErrorCode"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "code": {
                          "$ref": "#/components/schemas/Collections_GetItems_401ErrorCode"
                        }
                      }
                    }
                  ]
                }
              },
              "text/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "code": {
                          "$ref": "#/components/schemas/Collections_GetItems_401ErrorCode"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/Collections/user/{userId}": {
      "get": {
        "tags": [
          "Collections"
        ],
        "operationId": "Collections_GetUserCollections",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "take",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CollectionResponse"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CollectionResponse"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CollectionResponse"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "code": {
                          "$ref": "#/components/schemas/Collections_GetUserCollections_400ErrorCode"
                        }
                      }
                    }
                  ]
                }
              },
              "text/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "code": {
                          "$ref": "#/components/schemas/Collections_GetUserCollections_400ErrorCode"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "code": {
                          "$ref": "#/components/schemas/Collections_GetUserCollections_401ErrorCode"
                        }
                      }
                    }
                  ]
                }
              },
              "text/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "code": {
                          "$ref": "#/components/schemas/Collections_GetUserCollections_401ErrorCode"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CollectionResponse": {
        "required": [
          "id",
          "name",
          "description",
          "isPublic",
          "isSystem",
          "itemsCount",
          "createdAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "isPublic": {
            "type": "boolean"
          },
          "isSystem": {
            "type": "boolean"
          },
          "itemsCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": "integer",
            "format": "int32"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "Collections_Create_400ErrorCode": {
        "enum": [
          "General.Validation"
        ],
        "type": "string"
      },
      "Collections_Create_401ErrorCode": {
        "enum": [
          "Auth.Unauthorized"
        ],
        "type": "string"
      },
      "Collections_Create_409ErrorCode": {
        "enum": [
          "Collection.NameAlreadyExists"
        ],
        "type": "string"
      },
      "Collections_GetItems_400ErrorCode": {
        "enum": [
          "General.Validation"
        ],
        "type": "string"
      },
      "Collections_GetItems_401ErrorCode": {
        "enum": [
          "Auth.Unauthorized"
        ],
        "type": "string"
      },
      "Collections_GetItems_403ErrorCode": {
        "enum": [
          "Collection.Forbidden"
        ],
        "type": "string"
      },
      "Collections_GetItems_404ErrorCode": {
        "enum": [
          "Collection.NotFound"
        ],
        "type": "string"
      },
...
      "Collections_GetUserCollections_400ErrorCode": {
        "enum": [
          "General.Validation"
        ],
        "type": "string"
      },
      "Collections_GetUserCollections_401ErrorCode": {
        "enum": [
          "Auth.Unauthorized"
        ],
        "type": "string"
      },
...
      "Collections_Update_400ErrorCode": {
        "enum": [
          "General.Validation"
        ],
        "type": "string"
      },
      "Collections_Update_401ErrorCode": {
        "enum": [
          "Auth.Unauthorized"
        ],
        "type": "string"
      },
      "Collections_Update_403ErrorCode": {
        "enum": [
          "Collection.Forbidden"
        ],
        "type": "string"
      },
      "Collections_Update_404ErrorCode": {
        "enum": [
          "Collection.NotFound"
        ],
        "type": "string"
      },
      "Collections_Update_409ErrorCode": {
        "enum": [
          "Collection.NameAlreadyExists"
        ],
        "type": "string"
      },
...
      "CreateCollectionRequest": {
        "required": [
          "name",
          "description",
          "isPublic"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "isPublic": {
            "type": "boolean"
          }
        }
      },
...
      "PublicationResponse": {
        "required": [
          "id",
          "description",
          "imageUrl",
          "imageUrls",
          "createdAt",
          "userId",
          "userDisplayName",
          "tags",
          "clothes",
          "likesCount",
          "commentsCount",
          "assessmentsCount",
          "averageRating",
          "averageColorCoordination",
          "averageFitAndProportions",
          "averageOriginality",
          "averageOverallStyle",
          "isLikedByMe",
          "isSavedByMe"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "description": {
            "type": "string"
          },
          "imageUrl": {
            "type": "string"
          },
          "imageUrls": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "userDisplayName": {
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TagResponse"
            }
          },
          "clothes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ClothResponse"
            }
          },
          "likesCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": "integer",
            "format": "int32"
          },
          "commentsCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": "integer",
            "format": "int32"
          },
          "assessmentsCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": "integer",
            "format": "int32"
          },
          "averageRating": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": "number",
            "format": "double"
          },
          "averageColorCoordination": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": "number",
            "format": "double"
          },
          "averageFitAndProportions": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": "number",
            "format": "double"
          },
          "averageOriginality": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": "number",
            "format": "double"
          },
          "averageOverallStyle": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": "number",
            "format": "double"
          },
          "isLikedByMe": {
            "type": "boolean"
          },
          "isSavedByMe": {
            "type": "boolean"
          }
        }
      },
...
      "TagResponse": {
        "required": [
          "id",
          "name",
          "category"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "category": {
            "type": "string"
          }
        }
      },
...
      "UpdateCollectionRequest": {
        "required": [
          "name",
          "description",
          "isPublic"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "isPublic": {
            "type": "boolean"
          }
        }
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "JWT Authorization header using the Bearer scheme.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}